Restructure components and contexts, add uploader placeholder
