+++ components/PatientManager.js
