+++ components/UploaderSection.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const UploaderSection = () => {
+  return (
+    <section className="p-4 border rounded-md">
+      <h2 className="text-xl font-bold mb-4">File Uploader</h2>
+      {/* TODO: Implement file upload functionality */}
+      <p>Upload functionality coming soon.</p>
+    </section>
+  );
+};
+
+export default UploaderSection;
