+++ context/roleUtils.js
@@ -1,0 +1,4 @@
+export const checkUserRole = (user, roles = []) => {
+  const userRoles = user?.claims?.roles || [];
+  return roles.some(role => userRoles.includes(role));
+};
