+++ firestore.rules
@@ -1,10 +1,10 @@
 rules_version = '2';
 service cloud.firestore {
   match /databases/{database}/documents {
-    match /users/{userId} {
-      allow read, update, delete: if request.auth.uid == userId;
-      allow create: if request.auth != null;
-    }
+      match /users/{userId} {
+        allow read, update, delete: if request.auth.uid == userId;
+        allow create: if request.auth.uid == userId;
+      }
     
     match /users/{userId}/medicalCodingEntries/{entry} {
       allow read, write: if request.auth.uid == userId;
