+++ tailwind.config.js
@@ -10,11 +10,3 @@ module.exports = {
   },
   plugins: [],
 };
-
-// postcss.config.js
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
